+++ src/components/Heading/Heading.tsx
@@ -1,11 +1,13 @@
 import {
+  Box,
   chakra,
   Heading as ChakraHeading,
-  useStyleConfig,
+  useMultiStyleConfig,
 } from "@chakra-ui/react";
 import React, { forwardRef } from "react";
 
 import Link from "../Link/Link";
+import Text from "../Text/Text";
 import useNYPLBreakpoints from "../../hooks/useNYPLBreakpoints";
 
 export const headingSizesArray = [
@@ -58,6 +60,10 @@ export interface HeadingProps {
   size?: HeadingSizes;
   /** Optional prop used to remove default spacing */
   noSpace?: boolean;
+  /** String to populate the overline element */
+  overline?: string;
+  /** String to populate the subtitle element */
+  subtitle?: string;
   /** Inner text of the `<h*>` element */
   text?: string;
   /** Optional URL that header points to; when `url` prop is passed to
@@ -100,13 +106,15 @@ export const Heading = chakra(
         noSpace,
         size,
         text,
+        overline,
+        subtitle,
         url,
         urlClass,
         ...rest
       } = props;
       const finalLevel = getMappedLevel(level);
       const variant = size ? size : level;
-      const styles = useStyleConfig("Heading", {
+      const styles = useMultiStyleConfig("Heading", {
         variant,
         isCapitalized,
         isUppercase,
@@ -170,20 +178,60 @@ export const Heading = chakra(
             }
           : { fontSize: `mobile.heading.${finalRoot}${sizeIndex}` }
         : undefined;
-      return (
-        <ChakraHeading
-          as={asHeading}
-          className={className}
-          id={id}
-          ref={ref}
-          sx={{
-            ...styles,
-            ...responsiveStyles,
-          }}
-          {...rest}
-        >
-          {content}
-        </ChakraHeading>
+      const overlineSize = !isNaN(sizeIndex)
+        ? sizeIndex <= 2
+          ? "overline1"
+          : "overline2"
+        : undefined;
+      const subtitleSize = !isNaN(sizeIndex)
+        ? sizeIndex <= 2
+          ? "subtitle1"
+          : "subtitle2"
+        : undefined;
+      const wrapperStyles = styles.headingWrapper;
+      const headingStyles =
+        overline || subtitle
+          ? {
+              ...styles,
+              ...responsiveStyles,
+            }
+          : {
+              ...styles,
+              ...responsiveStyles,
+              ...wrapperStyles,
+            };
+      const finalContent = (
+        <>
+          {overline && (
+            <Text mb="xxs" size={overlineSize}>
+              {overline}
+            </Text>
+          )}
+          <ChakraHeading
+            as={asHeading}
+            className={className}
+            id={id}
+            ref={ref}
+            sx={{
+              ...headingStyles,
+            }}
+            {...rest}
+          >
+            {content}
+          </ChakraHeading>
+          {subtitle && (
+            <Text mt="xs" noSpace size={subtitleSize}>
+              {subtitle}
+            </Text>
+          )}
+        </>
+      );
+      return overline || subtitle ? (
+        <Box as="hgroup" sx={{ ...wrapperStyles }} {...rest}>
+          {finalContent}
+        </Box>
+      ) : (
+        <>{finalContent}</>
       );
     }
   )